modify user.js to be more modular
